+++ test/bones_api_mysql_test.dart
@@ -1,4 +1,6 @@
+@Tags(['docker'])
 @Timeout(Duration(seconds: 180))
+
 import 'package:bones_api/bones_api.dart';
 import 'package:bones_api/bones_api_adapter_mysql.dart';
 import 'package:docker_commander/docker_commander.dart';