+++ test/bones_api_version_test.dart
@@ -1,3 +1,4 @@
+@Tags(['version'])
 @TestOn('vm')
 
 import 'package:pubspec/pubspec.dart';